requirements.txt
matplotlib
keras
tensorflow-gpu
pandas
numpy
sklearn
hyperopt
lightgbm
